rails 4.2: fix migration methods + unhack things
fixes index_exists?, add_index, remove_index, and
 add_foreign_key method calls in migrations

unhacks remove_index to fail silently if the index
 doesn't exist (and removes some unnecessary remove_index calls)

also fixes migration specs for rails 4.2

refs #CNVS-26056

Reviewed-on: https://gerrit.instructure.com/70692
Tested-by: Jenkins
Product-Review: James Williams  <jamesw@instructure.com>
QA-Review: James Williams  <jamesw@instructure.com>

+++ db/migrate/20140806162559_drop_has_columns_from_context_external_tools.rb
@@ -20,7 +20,9 @@ class DropHasColumnsFromContextExternalTools < ActiveRecord::Migration
 
     EXTENSION_TYPES.each do |type|
       next if type == :homework_submission # note, there is no index for homework_submission
-      remove_index :context_external_tools, :"external_tools_#{type}"
+      if index_exists?(:context_external_tools, :"external_tools_#{type}")
+        remove_index :context_external_tools, :"external_tools_#{type}"
+      end
     end
   end
